Move CSV test data (#3044)
