+++ arrow-csv/test/data/uk_cities.csv
