+++ arrow-csv/test/data/uk_cities_with_headers.csv
