+++ arrow-csv/test/data/various_types.csv
