+++ arrow-csv/test/data/various_types_invalid.csv
