+++ arrow/Cargo.toml
@@ -267,4 +267,4 @@ required-features = ["test_utils", "ipc"]
 
 [[test]]
 name = "csv"
-required-features = ["csv"]
+required-features = ["csv", "chrono-tz"]
